refactor: make render fn separate
